requirements.txt
matplotlib
Pillow
tensorflow
tensorflow_datasets
